+++ CHANGELOG.md
@@ -33,6 +33,12 @@ This change log follows the conventions of
   network environments.
 - When disconnecting from a player’s dbserver, we politely send a
   teardown message and let it close the connection from its side.
+- A new utility method for creating a proxy around an object
+  which allows the underlying object to be garbage collected but
+  reloaded whenever the proxy is used. This is useful for building
+  user interfaces with lots of large objects which are mostly
+  off-screen, such as the track waveforms in the Beat Link Trigger
+  Show windows.
 
 ### Fixed
 
@@ -41,6 +47,9 @@ This change log follows the conventions of
   had always been broken, but was easy enough to fix. Now Beat Link
   Trigger can load tracks from rekordbox and rekordbox mobile onto
   players.
+- The various metadata finders were supposed to report the loss of
+  associated metadata when a player disappeared from the network, but
+  they were not doing so.
 
 ## [0.4.1] - 2018-10-28